Update de_DE.properties
Red Rocks is a slang word for heroin. I hope this fits the chemical properties in this mod, but calling it cocain (or red rock for that matter) would make no sense at all...
The "coke" translation could be wrong as well depending on the chemical structure in this mod. as the hydrochlorid version is just one possibilty to cut pure cocain. It is thos the mostly used variant and most likely what you will get when buying cocain on the street. (Dont ever do that!)